+++ homr/inference/transformer/tokenizer_rhythm.json
@@ -1,0 +1,103 @@
+{
+	"version": "1.0",
+	"truncation": null,
+	"padding": null,
+	"added_tokens": null,
+	"normalizer": null,
+	"pre_tokenizer": null,
+	"post_processor": null,
+	"decoder": null,
+	"model": {
+		"type": "WordLevel",
+		"vocab": {
+			"[PAD]": 0,
+			"[BOS]": 1,
+			"[EOS]": 2,
+			"+": 3,
+			"|": 4,
+			"barline": 5,
+			"clef-C1": 6,
+			"clef-C2": 7,
+			"clef-C3": 8,
+			"clef-C4": 9,
+			"clef-C5": 10,
+			"clef-F3": 11,
+			"clef-F4": 12,
+			"clef-F5": 13,
+			"clef-G1": 14,
+			"clef-G2": 15,
+			"keySignature-AM": 16,
+			"keySignature-AbM": 17,
+			"keySignature-BM": 18,
+			"keySignature-BbM": 19,
+			"keySignature-C#M": 20,
+			"keySignature-CM": 21,
+			"keySignature-CbM": 22,
+			"keySignature-DM": 23,
+			"keySignature-DbM": 24,
+			"keySignature-EM": 25,
+			"keySignature-EbM": 26,
+			"keySignature-F#M": 27,
+			"keySignature-FM": 28,
+			"keySignature-GM": 29,
+			"keySignature-GbM": 30,
+			"multirest-2": 31,
+			"multirest-3": 32,
+			"multirest-4": 33,
+			"multirest-5": 34,
+			"multirest-6": 35,
+			"multirest-7": 36,
+			"multirest-8": 37,
+			"multirest-9": 38,
+			"multirest-10": 39,
+			"note-breve": 40,
+			"note-breve.": 41,
+			"note-eighth": 42,
+			"note-eighth.": 43,
+			"note-half": 44,
+			"note-half.": 45,
+			"note-hundred_twenty_eighth": 46,
+			"note-long": 47,
+			"note-quarter": 48,
+			"note-quarter.": 49,
+			"note-sixteenth": 50,
+			"note-sixteenth.": 51,
+			"note-sixty_fourth": 52,
+			"note-sixty_fourth.": 53,
+			"note-thirty_second": 54,
+			"note-thirty_second.": 55,
+			"note-whole": 56,
+			"note-whole.": 57,
+			"rest-256th": 58,
+			"rest-512th": 59,
+			"rest-breve": 60,
+			"rest-eighth": 61,
+			"rest-eighth.": 62,
+			"rest-half": 63,
+			"rest-half.": 64,
+			"rest-hundred_twenty_eighth": 65,
+			"rest-long": 66,
+			"rest-quarter": 67,
+			"rest-quarter.": 68,
+			"rest-sixteenth": 69,
+			"rest-sixteenth.": 70,
+			"rest-sixty_fourth": 71,
+			"rest-sixty_fourth.": 72,
+			"rest-thirty_second": 73,
+			"rest-thirty_second.": 74,
+			"rest-whole": 75,
+			"rest-whole.": 76,
+			"timeSignature-/1": 77,
+			"timeSignature-/2": 78,
+			"timeSignature-/3": 79,
+			"timeSignature-/4": 80,
+			"timeSignature-/6": 81,
+			"timeSignature-/8": 82,
+			"timeSignature-/12": 83,
+			"timeSignature-/16": 84,
+			"timeSignature-/32": 85,
+			"timeSignature-/48": 86
+		},
+		"unk_token": "<unk>"
+	}
+}
