--- homr/transformer/tokenizer_pitch.json
@@ -1,86 +1,0 @@
-{
-	"version": "1.0",
-	"truncation": null,
-	"padding": null,
-	"normalizer": null,
-	"pre_tokenizer": null,
-	"post_processor": null,
-	"decoder": null,
-	"model": {
-		"type": "WordLevel",
-		"vocab": {
-			"nonote": 0,
-			"note-C0": 1,
-			"note-D0": 2,
-			"note-E0": 3,
-			"note-F0": 4,
-			"note-G0": 5,
-			"note-A0": 6,
-			"note-B0": 7,
-			"note-C1": 8,
-			"note-D1": 9,
-			"note-E1": 10,
-			"note-F1": 11,
-			"note-G1": 12,
-			"note-A1": 13,
-			"note-B1": 14,
-			"note-C2": 15,
-			"note-D2": 16,
-			"note-E2": 17,
-			"note-F2": 18,
-			"note-G2": 19,
-			"note-A2": 20,
-			"note-B2": 21,
-			"note-C3": 22,
-			"note-D3": 23,
-			"note-E3": 24,
-			"note-F3": 25,
-			"note-G3": 26,
-			"note-A3": 27,
-			"note-B3": 28,
-			"note-C4": 29,
-			"note-D4": 30,
-			"note-E4": 31,
-			"note-F4": 32,
-			"note-G4": 33,
-			"note-A4": 34,
-			"note-B4": 35,
-			"note-C5": 36,
-			"note-D5": 37,
-			"note-E5": 38,
-			"note-F5": 39,
-			"note-G5": 40,
-			"note-A5": 41,
-			"note-B5": 42,
-			"note-C6": 43,
-			"note-D6": 44,
-			"note-E6": 45,
-			"note-F6": 46,
-			"note-G6": 47,
-			"note-A6": 48,
-			"note-B6": 49,
-			"note-C7": 50,
-			"note-D7": 51,
-			"note-E7": 52,
-			"note-F7": 53,
-			"note-G7": 54,
-			"note-A7": 55,
-			"note-B7": 56,
-			"note-C8": 57,
-			"note-D8": 58,
-			"note-E8": 59,
-			"note-F8": 60,
-			"note-G8": 61,
-			"note-A8": 62,
-			"note-B8": 63,
-			"note-C9": 64,
-			"note-D9": 65,
-			"note-E9": 66,
-			"note-F9": 67,
-			"note-G9": 68,
-			"note-A9": 69,
-			"note-B9": 70
-		},
-		"unk_token": "<unk>"
-	}
-}
