+++ main.py
@@ -1,0 +1,3 @@
+from homr.main import main
+
+main()
